@@ -16,8 +16,8 @@
     "express": "^4.17.1",
     "morgan": "^1.9.1",
     "p5": "^0.8.0",
-    "react": "^16.4.0",
-    "react-dom": "^16.4.0",
+    "react": "^16.8.6",
+    "react-dom": "^16.8.6",
     "react-simple-keyboard": "^1.24.1"
   },
   "devDependencies": {
