@@ -1,60 +1,51 @@
-import React, {Component, Fragment} from 'react'
+import React, {useState, useEffect, Fragment} from 'react'
 import 'p5/lib/addons/p5.sound'
 import {AudioIn, SoundRecorder, SoundFile} from 'p5'
-
 const mic = new AudioIn()
 const recorder = new SoundRecorder()
 
-class Sampler extends Component {
-  state = {
-    micState: 0,
-    sample: new SoundFile()
-  }
+const Sampler = ({pressedKey, keyVal}) => {
+  const soundFile = new SoundFile()
+  const [micState, setMicState] = useState(0)
+  const [sample, setSample] = useState(soundFile)
 
-  componentDidMount() {
+  const enableMic = () => {
     mic.start()
     recorder.setInput(mic)
   }
 
-  // componentDidUpdate(prevProps) {
-  //   const {micState} = this.state
-  //   const {pressedKey, keyVal} = this.props
-  //   const {recordSample, playSample} = this
-  //   if(this.props !== prevProps && pressedKey === keyVal) {
-  //     if(micState === 1) playSample()
-  //     else recordSample()
-  //   }
-  // }
-
-  recordSample = evt => {
-    const {micState, sample} = this.state
-    if(micState === 0 && mic.enabled) {
-      recorder.record(sample)
-    } else if(micState === 1) {
-      recorder.stop()
-    }
-    const currentMicState = micState === 0 ? 1 : 0
-    this.setState({micState: currentMicState, sample})   
+  const onKeyDown = () => {
+    console.log(mic.getLevel())
+    window.addEventListener('keydown', ({key}) => {
+      if(key === keyVal) {
+        if(micState === 0) {
+          console.log('recording')
+          recorder.record(sample)
+          setMicState(1)
+        } else if(micState === 1) {
+          console.log('stopping')
+          recorder.stop()
+          setMicState(2)
+        }
+      } else if(key === keyVal.toUpperCase() && sample.buffer && !sample.isPlaying()) {
+        sample.play()
+        const emptySoundFile = new SoundFile()
+        setSample(emptySoundFile)
+        setMicState(0)
+      }
+    })
   }
+  useEffect(enableMic)
+  useEffect(onKeyDown)
 
-  playSample = evt => {
-    const {sample} = this.state
-    sample.play()
-  }
-
-  render() {
-    const {micState} = this.state
-    const {pressedKey, keyVal} = this.props
-    const {recordSample, playSample} = this
-    return (
-      <Fragment>
-        <div className='sampler w-150'>
-          <button className='record-btn w-150' onClick={recordSample}>Click to record sample</button>
-          <button className='play-btn w-150' onClick={playSample}>Click to play sample</button>
-        </div>
-      </Fragment>
-    )
-  }
+  return (
+    <Fragment>
+      <div className='sampler w-150'>
+        <h6 className='record-btn w-150'>Press {keyVal} to record sample</h6>
+        <button className='play-btn w-150' onClick={() => sample.play()} >Click to play sample</button>
+      </div>
+    </Fragment>
+  )
 }
 
 
